requirements.txt
Flask==0.8
Jinja2==2.6
Werkzeug==0.8.3
python-dateutil==2.1
six==1.1.0
wsgiref==0.1.2
gunicorn
